Add BVH acceleration structure

+++ src/rt_pathtracer/defs.h
@@ -6,8 +6,8 @@
 #include <limits>
 #include <cmath>
 
-const double INF = std::numeric_limits<double>::infinity();
-constexpr double EPS = std::numeric_limits<double>::epsilon();   // Numerical epsilon
+const double INF = std::numeric_limits<float>::infinity();
+constexpr double EPS = std::numeric_limits<float>::epsilon();   // Numerical epsilon
 const double DELTA = sqrt(EPS);                             // Delta to use
 
 #endif //PPGSO_DEFS_H
